shift pin and purge to admin module
